==> 01-Data-Full-Train.ipynb <==
from pymongo import MongoClient
from collections import defaultdict, Counter
from tqdm.notebook import tqdm
import networkx as nx
import random
import math
import pickle

client = MongoClient()
music = client['music_recommender']

def get_info_track(music, user):
    res = {}
    for e, t in enumerate(music.user_track_history.find_one({'username': user})['tracks']):
        t_t = (t['artist'], t['track'])
        res[t_t] = (e, t['timestamp'])
    return res

def load_graph(music, max_songs=None, ignore=False):
    users_count = music.user_scrobbles_history.count_documents({"tracks":{"$not":{"$size":0}}})
    track_count = music.track_info.count_documents({"spotify_id": {"$exists": True}})
    #Generate ids for the users
    print('Indexing users...')
    users = music.user_scrobbles_history.find({"tracks":{"$not":{"$size":0}}}, {"username": True})
    users = [u['username'] for u in tqdm(users, total=users_count) 
             if music.user_track_history.count_documents({'username': u['username']}) > 0]
    users_count = len(users)
    users_ids = {u: v for v, u in enumerate(users)}
    
    #Genereate ids for the tracks
    valid_spotify = {x['spotify_id'] for x in music.track_spotify_features.find()}
    print('Indexing songs...')
    val_artist = defaultdict(set)
    for u in music.user_track_history.find({}):
        for t in u['tracks']:
            val_artist[t['artist']].add(t['track'])
    #tracks = music.track_info.aggregate([{"$project": {"track": True, "artist": True}}])
    tracks = music.track_info.find({"spotify_id": {"$exists": True}},
                                   {"track": True, "artist": True, "_id": False, "spotify_id": True})
    artist_ids = defaultdict(dict)
    count = 0
    spotify = {}
    for e, t in tqdm(enumerate(tracks, start=len(users_ids)), total=track_count):
        if t['artist'] not in val_artist or t['track'] not in val_artist[t['artist']] or t['spotify_id'] not in valid_spotify:
            continue
        if t['spotify_id'] not in spotify:
            spotify[t['spotify_id']] = e
        artist_ids[t['artist']][t['track']] = spotify[t['spotify_id']]
        count += 1
    last_id = e 
    id_spotify = {e: s for s, e in spotify.items()}
    if max_songs is None:
        max_songs=count
    #Generate Graph
    print('Generating graph...')
    failed = 0
    g = nx.Graph()
    for u in tqdm(users):
        u = music.user_scrobbles_history.find_one({'username': u})
        info_tracks = get_info_track(music, u['username'])
        tracks = u['tracks']
        id_u = users_ids[u['username']]
        g.add_node(id_u)
        added = set()
        for i in range(min(max_songs, len(tracks))):
            t = tracks[i]
            if t['artist'] not in artist_ids or t['track'] not in artist_ids[t['artist']]:
                failed += 1
                if ignore:
                    continue
                last_id += 1
                artist_ids[t['artist']][t['track']] = last_id
            id_t = artist_ids[t['artist']][t['track']]
            if id_t in added: #Si esta agregada la actualiza
                g[id_u][id_t]['scrobbles'] += int(t['scrobbles'].replace(',', ''))
                if (t['artist'], t['track']) in info_tracks and g[id_u][id_t]['pos'] > info_tracks[(t['artist'], t['track'])][0]:
                    g[id_u][id_t]['pos'] = info_tracks[(t['artist'], t['track'])][0]
                    g[id_u][id_t]['date'] = info_tracks[(t['artist'], t['track'])][1]
                continue
            #Sino la agrego
            if (t['artist'], t['track']) not in info_tracks:
                g.add_edge(id_u, id_t, scrobbles=int(t['scrobbles'].replace(',', '')), 
                           pos=float('inf'),
                           date=None)
            else:
                g.add_edge(id_u, id_t, scrobbles=int(t['scrobbles'].replace(',', '')), 
                           pos=info_tracks[(t['artist'], t['track'])][0],
                           date=info_tracks[(t['artist'], t['track'])][1])
            added.add(id_t)
    print('Failed songs: {}'.format(failed))        
    return g, users_ids, artist_ids, spotify

g, users_ids, artist_ids, spotify_ids = load_graph(music, ignore=True)
#g, users_ids, artist_ids = load_graph(music)

def get_connected_users(g, users_ids):
    processed = set()
    components = list()
    users = set(users_ids.values())
    for u in tqdm(users):
        to_look = [u]
        component = set()
        while len(to_look) > 0:
            n = to_look.pop()
            if n in processed:
                continue
            processed.add(n)
            if n in users:
                component.add(n)
            to_look.extend(g[n])
        #Remove users without tracks
        non_track_users = set()
        for u in component:
            valid = False
            for n in g.neighbors(u):
                if n not in users:
                    valid = True
                    break
            if not valid:
                non_track_users.add(u)
        component = component - non_track_users
        #Remove users without tracks
        if len(component) > 0:
            components.append(component)
    components.sort(key=lambda x: len(x), reverse=True)
    return components

components = get_connected_users(g, users_ids)
for c in components:
    print(list(c)[:10])
    print((len(c) / len(users_ids)) * 100)

def sub_graph(g, users, users_ids, artist_ids):
    #Map to new user ids
    new_user_id = {u: e for e, u in enumerate(users)}
    n_users_ids = {u: new_user_id[n] for u, n in users_ids.items() if n in new_user_id}
    #Map to new artist-track ids
    tracks = set()
    for u in users:
        tracks.update(g[u])
    tracks = list(tracks)
    tracks.sort()
    new_track_id = {t: (e + len(n_users_ids)) for e, t in enumerate(tracks)}
    n_artist_ids = defaultdict(dict)
    for a, tracks in artist_ids.items():
        for t, i in tracks.items():
            if i in new_track_id:
                n_artist_ids[a][t] = new_track_id[i]
    #Generate Graph
    ng = nx.Graph()
    for u in users:
        nu = new_user_id[u]
        for t in g[u]:
            ng.add_edge(nu, new_track_id[t], 
                        scrobbles=g.edges[(u, t)]['scrobbles'],
                        pos=g.edges[(u, t)]['pos'], 
                        date=g.edges[(u, t)]['date'])
    return ng, n_users_ids, n_artist_ids, n_users_ids, new_track_id

sub_g, sub_users_ids, sub_artist_ids, _, track_map = sub_graph(g, components[0], users_ids, artist_ids)

len(sub_g.nodes()) - len(sub_users_ids)


len(sub_users_ids)

def transfer_tracks(g_or, g_dest, user, tracks):
    for t in tracks:
        g_dest.add_edge(user, t, 
                        scrobbles=g_or.edges[(user, t)]['scrobbles'], 
                        pos=g_or.edges[(user, t)]['pos'], 
                        date=g_or.edges[(user, t)]['date'])
    pass

'''
def train_test(g, users_ids, test_split=0.3):
    refs = Counter()
    #Count references to songs
    for u in users_ids.values():
        songs = set(g[u])
        for s in songs:
            refs[s] += 1
    #Generate graph
    g_train = nx.Graph()
    g_test = nx.Graph()
    non_songs_users = 0
    count_train = 0
    count_test = 0
    for u in tqdm(users_ids.values()):
        tracks = list(g[u])
        tracks.sort(key=lambda x: g[u][x]['pos'])
        split = min(math.ceil(len(tracks) * test_split), len(tracks) - 1)
        assert split >= 0
        for i in range(split):
            if refs[tracks[i]] == 1:
                split = i
                break
            refs[tracks[i]] -= 1
        if split == 0:
            print(u)
            non_songs_users += 1
        train = tracks[split:]
        test = tracks[:split]
        transfer_tracks(g, g_train, u, train)
        transfer_tracks(g, g_test, u, test)  
        count_train += len(train)
        count_test += len(test)
    print(f'Users without songs in testing {non_songs_users}')
    print(f'Train data: {count_train} Test data: {count_test}')
    return g_train, g_test  

'''
def train_test(g, users_ids, test_split=0.3):
    refs = Counter()
    #Count references to songs
    for u in users_ids.values():
        songs = set(g[u])
        for s in songs:
            refs[s] += 1
    #Generate graph
    g_train = nx.Graph()
    g_test = nx.Graph()
    non_songs_users = 0
    count_train = 0
    count_test = 0
    train_set = set()
    test_set = set()
    for u in tqdm(users_ids.values()):
        tracks = list(g[u])
        tracks.sort(key=lambda x: g[u][x]['pos'])
        v_len = len([x for x in tracks if g[u][x]['pos'] < float('inf')])
        #split = min(math.ceil(v_len * test_split), len(tracks) - 1)
        split = min(math.ceil(len(tracks) * test_split), len(tracks) - 1)
        assert split >= 0
        train = []
        test = []
        i = 0
        while i < len(tracks) and len(test) < split and g[u][tracks[i]]['pos'] < float('inf'):
            ct = tracks[i]
            if refs[ct] == 1:
                train.append(ct)
            else:
                test.append(ct)
                refs[ct] -= 1
            i += 1
        if len(test) == 0:
            #print(u)
            non_songs_users += 1
        train.extend(tracks[i:])
        transfer_tracks(g, g_train, u, train)
        transfer_tracks(g, g_test, u, test)  
        count_train += len(train)
        count_test += len(test)
        train_set.update(train)
        test_set.update(test)
    print(f'Users without songs in testing {non_songs_users}')
    print(f'Train data: {count_train} Test data: {count_test}')
    print(f'Train tracks data: {len(train_set)} Test tracks data: {len(test_set)}')
    return g_train, g_test 

g_train, g_test = train_test(sub_g, sub_users_ids)

unused_users = set()
for u in sub_users_ids.values():
    if u in g_test and len(list(g_test.neighbors(u))) == 0:
        unused_users.add(u)
    
print(len(unused_users))
g_test.remove_nodes_from(unused_users)

#Check all ids
not_used = [0] * len(sub_g.nodes())

for i in tqdm(sub_users_ids.values()):
    not_used[i] += 1

for tracks in tqdm(sub_artist_ids.values()):
    for i in tracks.values():
        not_used[i] += 1
        
total = 0
for e, i in enumerate(not_used):
    if i == 1:
        total += 1
    else:
        print(f'{e}: {i}')

print(total)
print(len(not_used))

t = list(sub_g[0].keys())[0]
i_track_map = {v: k for k, v in track_map.items()}
print(i_track_map[t])
ids_spotify = {v: k for k, v in spotify_ids.items()}
print(ids_spotify[i_track_map[t]])

spotify_ids[ids_spotify[i_track_map[t]]]

artist_ids['Black Kids']

sub_artist_ids['Black Kids']

with open('data/dataset.pickle', 'wb') as f:
    pickle.dump({'full': sub_g, 'train': g_train, 'test': g_test, 
                 'users': sub_users_ids, 'artist-tracks': sub_artist_ids}, f)

nx.draw(g)

nx.draw(g_train)

nx.draw(g_test)
