==> 02-Embeddings.ipynb <==
from pymongo import MongoClient
import pymongo

from tqdm.notebook import tqdm
import networkx as nx

import numpy as np
import os
import matplotlib.pyplot as plt
import pandas as pd

import pickle

with open('data/dataset.pickle', 'rb') as f:
    dataset = pickle.load(f)

data = []
g_train = dataset['train']
for u in tqdm(dataset['users'].values()):
    for s in g_train[u]:
        data.append([u, s, g_train.edges()[(u, s)]['scrobbles']])
 
df = pd.DataFrame(data=data, columns=['user', 'track', 'scrobbles'])

df.describe()

def search_superior(data, percentage=0.01, dmin=None, dmax=None):
    total = data.shape[0]
    if dmin is None:
        dmin = np.min(data)
    if dmax is None:
        dmax = np.max(data)
    mid = (dmax + dmin) // 2
    while dmin != mid and mid != dmax:
        p = np.sum(data >= mid) / total
        #print(f'{p} {dmin} {mid} {dmax}')
        if p > percentage:
            dmin = mid
        elif p < percentage:
            dmax = mid
        else:
            return mid
        mid = (dmax + dmin) // 2
    pmin = np.sum(data > dmin) / total
    pmax = np.sum(data > dmax) / total
    if abs(pmin - percentage) < abs(pmax - percentage):
        return dmin
    return dmax

print(search_superior(df['scrobbles'].values, percentage=0.01))
print(search_superior(df['scrobbles'].values, percentage=0.02))
print(search_superior(df['scrobbles'].values, percentage=0.03))
df['scrobbles'] = np.clip(df['scrobbles'], 0, search_superior(df['scrobbles'].values, percentage=0.03))

df.describe()

# Embeddings

x1 = np.asarray(df['user'], dtype=np.int32)[:, np.newaxis]
x2 = np.asarray(df['track'], dtype=np.int32)[:, np.newaxis] - len(dataset['users'])

import tensorflow as tf
from tensorflow.keras.layers import Input, Embedding, Dot, Reshape, Add, Lambda, Concatenate, Activation
from tensorflow.keras.models import Model
import functools
import tensorflow.keras.backend as K
from scipy.sparse import dok_matrix
import random


seed_value = 42
os.environ['PYTHONHASHSEED']=str(seed_value)
random.seed(seed_value)
np.random.seed(seed_value)
tf.random.set_seed(seed_value)
tf.compat.v1.set_random_seed(seed_value)

def base_model(users, tracks, vector_dim=64):
    """
    A Keras implementation of the GloVe architecture
    :param vocab_size: The number of distinct words
    :param vector_dim: The vector dimension of each word
    :return:
    """
    input_target = Input((1,), name='user_id')
    input_context = Input((1,), name='track_id')

    central_embedding = Embedding(users, vector_dim, input_length=1, name='user_emb')

    context_embedding = Embedding(tracks, vector_dim, input_length=1, name='track_emb')

    vector_target = central_embedding(input_target)
    vector_context = context_embedding(input_context)

    dot_product = Dot(axes=-1)([vector_target, vector_context])
    prediction = Reshape((1, ))(dot_product)
    prediction = Activation('sigmoid')(prediction)

    model = Model(inputs=[input_target, input_context], outputs=prediction)
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['binary_accuracy'])
    model.summary()

    return model

model = base_model(len(dataset['users']), len(dataset['train'].nodes())-len(dataset['users']), 10)

from tensorflow.keras.utils import Sequence
from sklearn.utils import shuffle
import math


class NegSamping(Sequence):
    
    def __init__(self, users, tracks, n_users, n_tracks, batch_size):
        self.users = users
        self.tracks = tracks
        self.n_users = n_users
        self.n_tracks = n_tracks
        self.batch_size = batch_size
        self.on_epoch_end()
    
    def __len__(self):
        return math.ceil(self.tracks.shape[0] / self.batch_size)
    
    def __getitem__(self, idx):
        x1 = self.users[idx * self.batch_size:(idx+1) * self.batch_size:, ...]
        x2 = self.tracks[idx * self.batch_size:(idx+1) * self.batch_size:, ...]
        y = np.ones((x1.shape[0], 1))
        x1_n = np.random.randint(0, self.n_users, size=x1.shape)
        x2_n = np.random.randint(0, self.n_tracks, size=x2.shape)
        y_n = np.zeros_like(y)
        return [np.concatenate((x1, x1_n), axis=0),
                np.concatenate((x2, x2_n), axis=0)], np.concatenate((y, y_n), axis=0)
    
    def on_epoch_end(self):
        self.users, self.tracks = shuffle(self.users, self.tracks)

ds = NegSamping(x1, x2, len(dataset['users']),len(dataset['train'].nodes())-len(dataset['users']), 500)

ds[0]

if not os.path.exists('data/embeddedings.h5'):
    model.fit(NegSamping(x1, x2,
                         len(dataset['users']), 
                         len(dataset['train'].nodes())-len(dataset['users']),
                         500), epochs=15) 
    model.save_weights('data/embeddedings.h5')
else:
    model.load_weights('data/embeddedings.h5')

print(model.layers[2].name)
users_embs = K.get_value(model.layers[2].embeddings)
print(model.layers[3].name)
tracks_embs = K.get_value(model.layers[3].embeddings)

model.predict([x1, x2], batch_size=10000)

del model

# Cosines

if not os.path.exists('data/cos.npz'):
    len_users = np.linalg.norm(users_embs, axis=1)[:, np.newaxis]
    len_tracks = np.linalg.norm(tracks_embs, axis=1)[:, np.newaxis]

    cos = (users_embs @ tracks_embs.T) / (len_users @ len_tracks.T)
    np.savez_compressed('data/cos.npz', cosines=cos)
else:
    cos = np.load('data/cos.npz')['cosines']
del users_embs
del tracks_embs

print(cos)
print(np.min(cos))
print(np.max(cos))
print(np.mean(cos))
print(np.std(cos))

# Communities test

from networkx.algorithms.community import greedy_modularity_communities

if not os.path.exists('data/commutities.pickle'):
    communities = list(greedy_modularity_communities(dataset['train']))
    with open('data/commutities.pickle', 'wb') as f:
        pickle.dump(communities, f)
else:
    with open('data/commutities.pickle', 'rb') as f:
        communities = pickle.load(f)

type(communities)

node_community = {}
for c, group in enumerate(communities):
    for u in group:
        node_community[u] = c

from collections import deque
c_sim = 0
c_diff = 0
c_all_cos = 0

for i in tqdm(range(len(dataset['users']))):
    for j in range(len(dataset['users']), len(dataset['train'].nodes())):
        c_all_cos += 1
        if node_community[i] == node_community[j]:
            c_sim += 1
        else:
            c_diff += 1
            
sim = np.empty((c_sim,))
diff = np.empty((c_diff,))
all_cos = np.empty((c_all_cos,))
i_sim = 0
i_diff = 0
i_all_cos = 0

for i in tqdm(range(len(dataset['users']))):
    for j in range(len(dataset['users']), len(dataset['train'].nodes())):
        current_cos = cos[i, j - len(dataset['users'])]
        all_cos[i_all_cos] = current_cos
        i_all_cos += 1
        if node_community[i] == node_community[j]:
            sim[i_sim] = current_cos
            i_sim += 1
        else:
            diff[i_diff] = current_cos
            i_diff += 1
            
print(c_diff == i_diff)
print(c_sim == i_sim)
print(c_all_cos == i_all_cos)

print(len(all_cos))
print(np.mean(all_cos))
print(np.std(all_cos))
print(len(sim))
print(np.mean(sim))
print(np.std(sim))
print(len(diff))
print(np.mean(diff))
print(np.std(diff))

from scipy.stats import ttest_ind

ttest_ind(sim, diff, alternative='greater', equal_var=False)

ttest_ind(sim, all_cos, alternative='greater', equal_var=False)

ttest_ind(all_cos, diff, alternative='greater', equal_var=False)
